Refactor CompatibilityView and AboutView
- Added a few things to the DeveloperView
- Rename some localizations and functions
- Put CompatibilityView into sections for more clarity
- Switch to isFlashAvailable from hasFlash
- Remove some unnecessary code, formatting changes
