+++ Malachite/Utilities/Preferences/MalachitePreferences.swift
@@ -7,7 +7,6 @@
 
 import Foundation
 import AppIntents
-// TODO: Rename a bunch of these preferences to be more concise in their meaning
 
 struct MalachitePreferences_AppContext: Codable, IntentResult {
     var value: Never?
