+++ Malachite/Views/AboutView/AboutView+Story.swift
@@ -1,0 +1,18 @@
+//
+//  AboutView+Story.swift
+//  Malachite
+//
+//  Created by Eva Isabella Luna on 8/16/25.
+//
+
+import SwiftUI
+
+extension AboutView {
+    struct Story: View {
+        var body: some View {
+            Section(header: Text("about.header.story")) {
+                Text("about.story")
+            }
+        }
+    }
+}
