+++ Malachite/Views/AboutView/AboutView.swift
@@ -1,0 +1,41 @@
+//
+//  AboutView.swift
+//  Malachite
+//
+//  Created by Eva Isabella Luna on 2/18/24.
+//
+
+import SwiftUI
+import GameKit
+
+struct AboutView: View {
+    /// A State variable used for determining whether or not this view is being presented as a modal.
+    var dismissAction: (() -> Void)
+    /// A variable to hold the existing instance of ``MalachiteClassesObject``.
+    var utilities = MalachiteClassesObject()
+    
+    /**
+     A variable used to hold the entire view.
+     
+     SwiftUI is weird...
+     Currently holds:
+     - Other variables to avoid type counting time issues.
+     - Handles initialization of variables required to show current settings.
+     - Navigation title of "About Malachite"
+     - Toolbar item for dismissing the view
+     */
+    var body: some View {
+        Form {
+            Info(utilities: utilities)
+            Story()
+            Credits(utilities: utilities)
+            Eggs(utilities: utilities) 
+        }
+        .navigationTitle("view.title.about")
+        .toolbar(content: {
+            ToolbarItemGroup(placement: .topBarTrailing) {
+                MalachiteToolbarUtils(action: self.dismissAction, image: "checkmark", primary: true)
+            }
+        })
+    }
+}
