+++ Malachite/Views/QuickHelpView/QuickHelpView+Photo.swift
@@ -12,9 +12,9 @@ extension QuickHelpView {
         /// A variable to hold the photo settings section.
         var body: some View {
             Section(header: Text("settings.header.photo"), footer: Text("settings.footer.photo")) {
-                Builder(title: Text("settings.option.photo.fileformat"), subtitle: Text("settings.detail.photo.fileformat")) {}
-                Builder(title: Text("settings.option.photo.hdr"), subtitle: Text("settings.detail.photo.hdr")) {}
-                Builder(title: Text("settings.option.photo.continuous"), subtitle: Text("settings.detail.photo.continuous")) {}
+                createQuickHelpRow(title: Text("settings.option.photo.fileformat"), subtitle: Text("settings.detail.photo.fileformat")) 
+                createQuickHelpRow(title: Text("settings.option.photo.hdr"), subtitle: Text("settings.detail.photo.hdr"))
+                createQuickHelpRow(title: Text("settings.option.photo.continuous"), subtitle: Text("settings.detail.photo.continuous"))
             }
         }
     }
