+++ Malachite/Views/QuickHelpView/QuickHelpView+Preview.swift
@@ -12,9 +12,9 @@ extension QuickHelpView {
         /// A variable to hold the preview settings section.
         var body: some View {
             Section(header: Text("settings.header.preview"), footer: Text("settings.footer.preview")) {
-                Builder(title: Text("settings.option.preview.aspect_ratio"), subtitle: Text("settings.detail.preview.aspect_ratio")) {}
-                Builder(title: Text("settings.option.preview.sbtlz"), subtitle: Text("settings.detail.preview.sbtlz")) {}
-                Builder(title: Text("settings.option.preview.zoom_maximum"), subtitle: Text("settings.detail.preview.zoom_maximum")) {}
+                createQuickHelpRow(title: Text("settings.option.preview.aspect_ratio"), subtitle: Text("settings.detail.preview.aspect_ratio")) 
+                createQuickHelpRow(title: Text("settings.option.preview.sbtlz"), subtitle: Text("settings.detail.preview.sbtlz"))
+                createQuickHelpRow(title: Text("settings.option.preview.zoom_maximum"), subtitle: Text("settings.detail.preview.zoom_maximum"))
             }
         }
     }
