+++ MalachiteWatch/ContentView.swift
@@ -10,17 +10,10 @@ import SwiftUI
 struct ContentView: View {
     var body: some View {
         VStack {
-            if #available(watchOS 8.0, *) {
-                Image(systemName: "globe")
-                    .imageScale(.large)
-                    .foregroundStyle(.tint)
-                Text("Hello, world!")
-            } else {
-                Image(systemName: "globe")
-                    .imageScale(.large)
-                    .foregroundColor(.accentColor)
-                Text("Hello, world!")
-            }
+            Image(systemName: "globe")
+                .imageScale(.large)
+                .foregroundStyle(.tint)
+            Text("Hello, world!")
         }
         .padding()
     }
